refactor: extract logic + move files to subfolder
