+++ packages/shared/lib/core/wallet/actions/index.ts
@@ -1,13 +1,10 @@
 export * from './claimActivity'
 export * from './getAccountAssetsForSelectedAccount'
 export * from './rejectActivity'
-export * from './linkActivityAndClaimingTransaction'
-export * from './loadAccountActivities'
-export * from './hideActivitiesForFoundries'
 export * from './refreshAccountAssetsForActiveProfile'
 export * from './requestAndStorePersistedAsset'
 export * from './sendOutput'
 export * from './setAsyncActivitiesToClaimed'
-export * from './setAsyncStatusOfAccountActivities'
 export * from './mintNativeToken'
 export * from './tryGetAndStoreAssetFromPersistedAssets'
+export * from './activities'
