+++ packages/shared/lib/core/wallet/interfaces/index.ts
@@ -3,7 +3,7 @@ export * from './activity.interface'
 export * from './asset-balance.interface'
 export * from './asset.interface'
 export * from './base-token-metadata.interface'
-export * from './processed-output.interface'
+export * from './processed-transaction.interface'
 export * from './claimed-activities.interface'
 export * from './filter.interface'
 export * from './hidden-activities.interface'
